+++ backend/src/routes/auth/handlers/edit.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance, FastifyRequest } from "fastify";
 import type { Admin } from "@viernulvier/shared/index.js";
 import { AdminSchema } from "@viernulvier/shared/index.js";
-import { getMetadata, getParam, parseFirstRow, safeParse } from "@/routes/helpers.js";
+import { getMetadata, getParam, parseFirstRow, parse } from "@/routes/helpers.js";
 import { z } from "zod";
 import { hashPassword } from "./hash.js";
 
@@ -18,7 +18,7 @@ const EditAdminBodySchema = AdminSchema.pick({ username: true }).extend({
  */
 export async function editAdmin(server: FastifyInstance, request: FastifyRequest): Promise<Admin | null> {
   const id = getParam(request, "id");
-  const body = safeParse(server, EditAdminBodySchema, request.body);
+  const body = parse(server, EditAdminBodySchema, request.body);
 
   const { admin, current_time } = getMetadata();
