+++ backend/src/routes/auth/handlers/fetch.ts
