+++ backend/src/routes/auth/handlers/hash.ts
