+++ backend/src/routes/auth/handlers/index.ts
@@ -1,0 +1,5 @@
+export * from "./fetch.js";
+export * from "./create.js";
+export * from "./edit.js";
+export * from "./replace.js";
+export * from "./delete.js";
