+++ shared/src/types/admin.ts
@@ -4,9 +4,9 @@ import { primaryKey } from "./helpers.js";
 import { createSchema } from "./index.js";
 
 export const AdminBase = {
-  id: primaryKey(),
-  username: z.string().max(32),
-  profile_picture: z.url().max(2048).nullable(),
+  id: primaryKey().describe("Primary key of admin."),
+  username: z.string().max(32).describe("Username of admin."),
+  profile_picture: z.url().max(2048).nullable().describe("Profile picture (url) of admin."),
 };
 
 export const AdminSchema = createSchema(AdminBase);
